# backendpython/docs/openapi.yaml
...
openapi: 3.0.0
info:
  title: WikiCap API
  description: API for year-based cultural and historical data
  version: 1.0.0
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'

## Här under "paths" definieras alla endpoints och deras funktioner. Det är bara att följa strukturen som den är nu för varje ny endpoint.
## Tänk på indenteringen då detta är i princip som JSON fast utan "{" och "}" där indenteringen gör jobbet istället.
paths:
...
  /api/v1/year/{year}/nobel:
    get:
      summary: Get Nobel Prize laureates for a specific year
      description: Returns Nobel Prize laureates for the specified year, grouped by category (name, motivation, and image URL when available).
      tags:
        - Nobel
      parameters:
        - name: year
          in: path
          required: true
          description: The year to retrieve Nobel Prize laureates for
          schema:
            type: integer
            minimum: 1900
            maximum: 2100
          example: 2019
      responses:
        '200':
          description: Nobel Prize laureates successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NobelResponse'
        '400':
          description: Invalid year parameter
        '404':
          description: No Nobel Prize data found for the specified year
        '429':
          description: Rate limit exceeded when accessing Nobel Prize data
        '500':
          description: Internal server error
...
  /api/v1/year/{year}/songs:
    get:
      summary: Get popular songs for a specific year from Spotify
      description: Returns popular songs for the specified year, ranked by popularity from Spotify
        and based on limit how many songs to retrieve.
      tags:
        - Songs
      parameters:
        - name: year
          in: path
          required: true
          description: The year to retrieve songs for
          schema:
            type: integer
            minimum: 1900
            maximum: 2026
          example: 1987
      responses:
        '200':
          description: Top songs successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongsResponse'
        '400':
          description: Invalid year parameter
        '404':
          description: No songs found for the specified year
        '500':
          description: Internal server error
        '502':
          description: Bad gateway error when accessing Spotify
        '503':
          description: Spotify se.wikicap.service is currently unavailable. Please try again later.
components:
  schemas:
...
    NobelResponse:
      type: object
      properties:
        year:
          type: integer
          description: The requested year
          example: 2019
        nobel_prizes:
          $ref: '#/components/schemas/NobelPrizes'
    NobelPrizes:
      type: object
      description: Nobel Prize categories mapped to a list of laureates
      additionalProperties:
        type: array
        items:
          $ref: '#/components/schemas/Laureate'
    Laureate:
      type: object
      properties:
        name:
          type: string
          description: Laureate name
          example: "Olga Tokarczuk"
        motivation:
          type: string
          nullable: true
          description: Prize motivation/justification (if available)
          example: "for a narrative imagination..."
        image:
          type: string
          nullable: true
          description: Laureate image URL (if available)
          example: "https://upload.wikimedia.org/wikipedia/commons/.../Olga_Tokarczuk.jpg"
...
    SongsResponse:
      type: object
      properties:
        year:
          type: integer
          description: The requested year
          example: 1987
        top_songs:
          type: array
          description: List of popular songs for the year
          items:
            $ref: '#/components/schemas/Song'
        source:
          type: string
          description: Data source
          example: Spotify
    Song:
      type: object
      properties:
        title:
          type: string
          description: Title of the song
          example: 'In the end'
        artist:
          type: string
          description: Artist name
          example: 'Linkin Park'
        album:
          type: string
          description: Album name
          example: 'Hybrid Theory (Bonus Edition)'
        release_date:
          type: string
          format: date
          description: Release date
          example: '2000-10-24'
        spotify_url:
          type: string
          format: uri
          description: URL to the song on Spotify
          example: 'https://open.spotify.com/track/xyz...'
        image:
          type: string
          format: uri
          description: URL to the album cover image
          example: 'https://i.scdn.co/image/abc123...'